ensure we emit a rename command for renamed files

+++ src/libsync/propagateremotemove.cpp
@@ -19,6 +19,7 @@
 #include "common/syncjournalfilerecord.h"
 #include "filesystem.h"
 #include "common/asserts.h"
+#include <QFileInfo>
 #include <QFile>
 #include <QStringList>
 #include <QDir>
@@ -79,7 +80,7 @@ void PropagateRemoteMove::start()
         return;
 
     QString origin = propagator()->adjustRenamedPath(_item->_file);
-    qCDebug(lcPropagateRemoteMove) << origin << _item->_renameTarget;
+    qCInfo(lcPropagateRemoteMove) << origin << _item->_renameTarget;
 
     QString targetFile(propagator()->fullLocalPath(_item->_renameTarget));
 
@@ -237,10 +238,14 @@ void PropagateRemoteMove::finalize()
     auto &vfs = propagator()->syncOptions()._vfs;
     auto pinState = vfs->pinState(_item->_originalFile);
 
-    // Delete old db data.
-    propagator()->_journal->deleteFileRecord(_item->_originalFile);
-    if (!vfs->setPinState(_item->_originalFile, PinState::Inherited)) {
-        qCWarning(lcPropagateRemoteMove) << "Could not set pin state of" << _item->_originalFile << "to inherited";
+    const auto targetFile = propagator()->fullLocalPath(_item->_renameTarget);
+
+    if (QFileInfo::exists(targetFile)) {
+        // Delete old db data.
+        propagator()->_journal->deleteFileRecord(_item->_originalFile);
+        if (!vfs->setPinState(_item->_originalFile, PinState::Inherited)) {
+            qCWarning(lcPropagateRemoteMove) << "Could not set pin state of" << _item->_originalFile << "to inherited";
+        }
     }
 
     SyncFileItem newItem(*_item);
@@ -254,6 +259,13 @@ void PropagateRemoteMove::finalize()
             newItem._size = oldRecord._fileSize;
         }
     }
+
+    if (!QFileInfo::exists(targetFile)) {
+        propagator()->_journal->commit("Remote Rename");
+        done(SyncFileItem::Success);
+        return;
+    }
+
     const auto result = propagator()->updateMetadata(newItem);
     if (!result) {
         done(SyncFileItem::FatalError, tr("Error updating metadata: %1").arg(result.error()));
